hikey: Add fstab overlay for verity on vendor.
Add a second android fstab overlay which can be used with the
combination of system-as-root and verity. On older builds, with treble,
but not system-as-root or verity, we can continue to use the older fstab
definition. The older fstab definition can be removed later, when using
older userspace is no longer required.

Bug: None.
Test: Boot hikey with system and vendor verity enabled

Change-Id: Iab64f3fffe9ccde0c61f9fac17b145c3175563be
Signed-off-by: Sandeep Patil <sspatil@google.com>