arm64: dts: hi3660: fix adv7533 interrupt
As reported here:
https://bugs.96boards.org/show_bug.cgi?id=830

The adv7533 chip which is used to transmit a video signal to
HDMI from MIPI/DSI has the incorrect interrupt in the device
tree. It looks like this was inadvertently duplicated from the
original Hikey.

According to the device tree (hi3660-hikey960.dts),
GPIO_180_HDMI_INT is in the gpio22 block, and is the fifth pin
(pin #4). When using Debian and upstream kernel, HDMI hotplug
did not work. With this, hotplug appears to work correctly now.

[jstultz: Added more context from the bug]
Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I240b5515ca15c5df4cefa9fdfbfee23bb3422cde