MALI: Avoid null pointer deref
So the utgard driver for hikey hasn't ever really been
configured to actually use the dmabuf fences.

This results in the sync fence to not properly have
the fence ops configured, and due to changes in the
userland usage, we're now hitting null pointer
dereferences.

I took a brief swing at trying to properly enable
CONFIG_MALI_DMA_BUF_FENCE=y and reworking the use
of the old reservation_objects but even after doing
that I was still hitting crashes in mali_timeline.c

So for now, lets just skip the fence call here if it
is null and avoid the crash.

Change-Id: Ic8b310e0e331ebed584a0517e1bc09a44ff9fe60
Signed-off-by: John Stultz <john.stultz@linaro.org>